[SOCA-466] Deprecate meter scan modes (#24)
* Use 'Auto' Meter scan mode in Analog / Digital / Dot Matrix examples

This makes the JSON configs for examples using the meter plugin change to the scan mode 'Auto' when previously among the following: "Digital", "Analog", and "Dot Matrix" (now marked as deprecated).

Reason: Enhancements to the training models has reached the point where 'Auto' script can already be used in place of the other three modes mentioned. Furthermore, the script files corresponding to the deprecated modes have since been removed from the resources bundle, so in order to avoid the scanner failing when attempting to load them, they should (until they themselves are removed) point to the Auto script file.